TC v3 (title_classifier) · v2.9.1 hotfix — set_enum smallint param cast
Live bug: async_set_enum reused $2 both as the smallint `enum` value and in
`CASE WHEN $2 <> 0` (integer literal). asyncpg deduced conflicting types
(integer vs smallint) → AmbiguousParameterError, so EVERY v3 set_enum failed
against real Postgres. This was also the real root cause of the earlier
"enum springs back" (the DB write silently failed; the React panel now surfaces
the error instead of reverting).

Fix: cast the param to ::smallint in both uses. Guard test asserts the cast is
present (FakePool can't reproduce PG type inference).

manifest 2.9.0 → 2.9.1. Backend-only; frontend unchanged. py_compile + pytest 175.

+++ custom_components/title_classifier/store_v3.py
@@ -221,8 +221,8 @@ class CatalogStoreV3:
         row = await self._pool.fetchrow(
             f"""
             UPDATE tc_v3_catalog
-               SET enum = $2,
-                   hidden_at = CASE WHEN $2 <> 0 THEN NULL ELSE hidden_at END,
+               SET enum = $2::smallint,
+                   hidden_at = CASE WHEN $2::smallint <> 0 THEN NULL ELSE hidden_at END,
                    updated_by = $3, updated_at = now()
              WHERE id = $1
             RETURNING {_CAT_COLUMNS}
